third_party/WebKit/LayoutTests/http/tests/permissions/iframe-geolocation-permissions-expected.txt

CONSOLE WARNING: line 15: A cross-origin iframe needs its permissions attribute properly set in order to use the geolocation API.
